+++ nym-vpn-core/crates/nym-vpn-lib-types/src/lib.rs
@@ -89,9 +89,9 @@ pub use gateway::{
 pub use gateway_independence::GatewayIndependence;
 pub use gateway_selection_algorithm::GatewaySelectionAlgorithmConfig;
 pub use network::{
-    ApiUrl, ChainDetails, DenomDetailsOwned, FeatureFlags, FlagValue, Network,
-    NetworkCompatibility, NymContracts, NymNetworkDetails, NymVpnNetwork, ParsedAccountLinks,
-    SystemConfiguration, SystemMessage, ValidatorDetails,
+    ApiUrl, ChainDetails, DenomDetailsOwned, DnsFallback, FeatureFlags, FlagValue, Network,
+    NetworkCompatibility, NymContracts, NymNetworkDetails, NymNetworkingSpecifics, NymVpnNetwork,
+    ParsedAccountLinks, SystemConfiguration, SystemMessage, ValidatorDetails,
 };
 pub use network_stats::{NetworkStatisticsConfig, NetworkStatisticsIdentity};
 pub use paths::LogPath;